sync

+++ .install.sh
@@ -12,22 +12,23 @@ if ! command -v nix >/dev/null; then
     read -r REPLY
     if [ "$REPLY" = "y" ] || [ "$REPLY" = "Y" ]; then
         if [[ "$OSTYPE" == "darwin"* ]]; then
-            curl --proto '=https' --tlsv1.2 -L https://nixos.org/nix/install | sh
+            curl --proto '=https' --tlsv1.2 -L 'https://nixos.org/nix/install' | sh
         else
-            curl --proto '=https' --tlsv1.2 -L https://nixos.org/nix/install | sh -s -- --no-daemon
+            curl --proto '=https' --tlsv1.2 -L 'https://nixos.org/nix/install' | sh -s -- --no-daemon
         fi
-        [ -f "$HOME/.nix-profile/etc/profile.d/nix.sh" ] && source "$HOME/.nix-profile/etc/profile.d/nix.sh"
+        [ -f "/nix/var/nix/profiles/default/etc/profile.d/nix-daemon.sh" ] && source "/nix/var/nix/profiles/default/etc/profile.d/nix-daemon.sh"
         [ -f "/nix/var/nix/profiles/default/etc/profile.d/nix.sh" ] && source "/nix/var/nix/profiles/default/etc/profile.d/nix.sh"
-    else
-        return
+        [ -f "$HOME/.nix-profile/etc/profile.d/nix.sh" ] && source "$HOME/.nix-profile/etc/profile.d/nix.sh"
     fi
+fi
 
+if command -v nix >/dev/null; then
     printf "Install nix packages? (y/N): "
     read -r REPLY
     if [ "$REPLY" = "y" ] || [ "$REPLY" = "Y" ]; then
         mkdir -p "$HOME/.config/nixpkgs/"
-        curl --silent https://gitlab.com/blakejc/dotfiles/-/blob/main/.config/nixpkgs/packages.nix?ref_type=heads > "$HOME/.config/nixpkgs/packages.nix"
-        curl --silent https://gitlab.com/blakejc/dotfiles/-/blob/main/.config/nixpkgs/config.nix?ref_type=heads > "$HOME/.config/nixpkgs/config.nix"
+        curl --silent 'https://gitlab.com/blakejc/dotfiles/-/raw/main/.config/nixpkgs/config.nix?ref_type=heads' > "$HOME/.config/nixpkgs/config.nix"
+        curl --silent 'https://gitlab.com/blakejc/dotfiles/-/raw/main/.config/nixpkgs/packages.nix?ref_type=heads' > "$HOME/.config/nixpkgs/packages.nix"
         nix-channel --update
         nix-env -f "$HOME/.config/nixpkgs/packages.nix" -i
         rm -r "$HOME/.config/nixpkgs/"
@@ -42,9 +43,8 @@ if [[ "$OSTYPE" == "darwin"* ]]; then
         printf "Install brew? (y/N): "
         read -r REPLY
         if [ "$REPLY" = "y" ] || [ "$REPLY" = "Y" ]; then
-            /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
+            /bin/bash -c "$(curl -fsSL 'https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh')"
             [ -f /opt/homebrew/bin/brew ] && eval "$(/opt/homebrew/bin/brew shellenv)"
-
         fi
     fi
 
@@ -60,13 +60,13 @@ if [[ "$OSTYPE" == "darwin"* ]]; then
                 else
                     brew upgrade --cask "${item}"
                 fi
-            done < <(curl --silent https://gitlab.com/blakejc/dotfiles/-/raw/main/.listcask?ref_type=heads)
+            done < <(curl --silent 'https://gitlab.com/blakejc/dotfiles/-/raw/main/.listcask?ref_type=heads')
         fi
     fi
 fi
 
 if command -v git >/dev/null; then
-    bash <(curl --silent https://gitlab.com/blakejc/dotfiles/-/blob/main/.gitalias/key?ref_type=heads)
+    bash <(curl --silent 'https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh')
     git clone --bare git@gitlab.com:blakejc/dotfiles.git "$HOME/.dotfiles"
     if command -v just >/dev/null; then
         just deploy-dotfiles || just deploy-dotfiles-safe
